Finally a good, pinned build
This has been struggling to build for quite some time. For many rounds
the git checkouts were not succeeding. Then, the git repositories were
not matching what used to be.
It would be a really good idea to figure out all the dependencies that
are needing to be packaged, but I'm not sure I have the gumption to do
that at this point